pyproject.toml
[project]
name = "ff-intelligent-mvp"
version = "1.0.0"
description = "FFmpeg batch processing desktop tool"
requires-python = ">=3.10"
dependencies = [
    "pywebview>=6.0",
    "static-ffmpeg",
    "loguru>=0.7",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pywebvue", "core"]
